+++ server/plugins/runtime-config-validation.ts
@@ -1,3 +1,19 @@
+/*
+ * Copyright (C) 2026 Kolja Nolte
+ * https://meow.yanawa.io
+ * info@meow.yanawa.io
+ *
+ * This work is licensed under the MIT License. You are free to use, modify,
+ * and distribute this work, provided that you include the copyright notice
+ * and this permission notice in all copies or substantial portions of the work.
+ * For more information, visit: https://opensource.org/licenses/MIT
+ *
+ * @author    Kolja Nolte
+ * @license   MIT
+ * @date      2026
+ * @website   https://meow.yanawa.io
+ */
+
 import { assertProductionRuntimeSecrets } from '../utils/runtimeConfigValidation'
 
 export default defineNitroPlugin(() => {